even better controller for player
i also added the eye animations properly. I still need to do testing for enemies before i can propely implement the attacks, but i am making it modular for both greany and purpley to work with the same character controller. Do know one set of eye animations you cannot see yet though, that being the "hit" animation, though setting the PlayerHit bool in the ThirdPersonMovement script to true while playmode is on so that you can see it, asside from that i am going to quickly get the animations for the attack done and it's attack trigger, and then i will finally add the enemies!